chore: BuzzerTable.jsx -> .tsx
